fix(config): 🐛 fix migration setup after regression
The refactor of splitting models into multiple files caused a regression
in the alembic migration setup, where the models were not imported
correctly in the env.py file. This commit fixes the imports, through the
models/__init__.py file in the models directory.

Side Effect: render as batch mode for improved ENUM handling